Integrate new bluetooth connection management #5 - Support Frame Device
