+++ app/lib/providers/onboarding_provider.dart
@@ -212,7 +212,7 @@ class OnboardingProvider extends BaseProvider with MessageNotifierMixin implemen
 
     ServiceManager.instance().device.subscribe(this, this);
 
-    _findDevicesTimer = Timer.periodic(const Duration(seconds: 2), (timer) async {
+    _findDevicesTimer = Timer.periodic(const Duration(seconds: 4), (timer) async {
       ServiceManager.instance().device.discover();
     });
   }
